Split out logging to its own module in rascaline-c-api
